lola: Extract skeleton shared memory functionality
The Skeleton currently performs Skeleton specific functionality (e.g.
PrepareOffer, PrepareStopOffer(), managing instance / usage marker
files, managing method subscription etc.). It also performs all
functionality to create / open shared memory and also insert events into
the maps in shared memory.

This commit splits up the functionality so that we have a single class
which is responsible for interacting with shared memory. This makes the
code much easier to navigate since the old skeleton code was a huge
monolith and also will allow us to optionally split up our unit tests as
well into skeleton functionality and shared memory operations. We could
even introduce an interface for SkeletonMemoryManager to completely
decouple the skeleton functionality tests from shared memory.

For now, we treat Skeleton and SkeletonMemoryManager as a single unit,
so we don't test the public interface of SkeletonMemoryManager.
